mcp-toolbox/planner-agents/requirements.txt
google-adk
toolbox-core
google-adk[eval]